+++ apps/backend/backend_client.c
@@ -219,10 +219,20 @@ from_client_get_config(clicon_handle h,
 	clicon_err(OE_XML, 0, "db not found");
 	goto done;
     }
+    if (xmldb_validate_db(db) < 0){
+	cprintf(cbret, "<rpc-reply><rpc-error>"
+		"<error-tag>invalid-value</error-tag>"
+		"<error-type>protocol</error-type>"
+		"<error-severity>error</error-severity>"
+		"<error-message>No such database: %s</error-message>"
+		"</rpc-error></rpc-reply>", db);
+	goto ok;
+    }
+
     if ((xfilter = xml_find(xe, "filter")) != NULL)
 	if ((selector = xml_find_value(xfilter, "select"))==NULL)
 	    selector="/";
-    if (xmldb_get(h, db, selector, &xret, NULL, NULL) < 0){
+    if (xmldb_get(h, db, selector, &xret) < 0){
 	cprintf(cbret, "<rpc-reply><rpc-error>"
 		"<error-tag>operation-failed</error-tag>"
 		"<error-type>application</error-type>"
@@ -276,6 +286,16 @@ from_client_edit_config(clicon_handle h,
 	clicon_err(OE_XML, 0, "db not found");
 	goto done;
     }
+    if (xmldb_validate_db(target) < 0){
+	cprintf(cbret, "<rpc-reply><rpc-error>"
+		"<error-tag>invalid-value</error-tag>"
+		"<error-type>protocol</error-type>"
+		"<error-severity>error</error-severity>"
+		"<error-message>No such database: %s</error-message>"
+		"</rpc-error></rpc-reply>", target);
+	goto ok;
+    }
+
     /* Check if target locked by other client */
     piddb = xmldb_islocked(h, target);
     if (piddb && mypid != piddb){
@@ -356,6 +376,16 @@ from_client_lock(clicon_handle h,
 		"</rpc-error></rpc-reply>");
 	goto ok;
     }
+    if (xmldb_validate_db(db) < 0){
+	cprintf(cbret, "<rpc-reply><rpc-error>"
+		"<error-tag>invalid-value</error-tag>"
+		"<error-type>protocol</error-type>"
+		"<error-severity>error</error-severity>"
+		"<error-message>No such database: %s</error-message>"
+		"</rpc-error></rpc-reply>", db);
+	goto ok;
+    }
+
     /*
      * A lock MUST not be granted if either of the following conditions is true:
      * 1) A lock is already held by any NETCONF session or another entity.
@@ -408,6 +438,15 @@ from_client_unlock(clicon_handle h,
 		"<error-severity>error</error-severity>"
 		"<error-info><bad-element>target</bad-element></error-info>"
 		"</rpc-error></rpc-reply>");
+	goto ok;
+    }
+    if (xmldb_validate_db(db) < 0){
+	cprintf(cbret, "<rpc-reply><rpc-error>"
+		"<error-tag>invalid-value</error-tag>"
+		"<error-type>protocol</error-type>"
+		"<error-severity>error</error-severity>"
+		"<error-message>No such database: %s</error-message>"
+		"</rpc-error></rpc-reply>", db);
 	goto ok;
     }
     piddb = xmldb_islocked(h, db);
@@ -534,6 +573,16 @@ from_client_copy_config(clicon_handle h,
 		"</rpc-error></rpc-reply>");
 	goto ok;
     }
+    if (xmldb_validate_db(source) < 0){
+	cprintf(cbret, "<rpc-reply><rpc-error>"
+		"<error-tag>invalid-value</error-tag>"
+		"<error-type>protocol</error-type>"
+		"<error-severity>error</error-severity>"
+		"<error-message>No such database: %s</error-message>"
+		"</rpc-error></rpc-reply>", source);
+	goto ok;
+    }
+
     if ((target = netconf_db_find(xe, "target")) == NULL){
 	cprintf(cbret, "<rpc-reply><rpc-error>"
 		"<error-tag>missing-element</error-tag>"
@@ -541,6 +590,15 @@ from_client_copy_config(clicon_handle h,
 		"<error-severity>error</error-severity>"
 		"<error-info><bad-element>target</bad-element></error-info>"
 		"</rpc-error></rpc-reply>");
+	goto ok;
+    }
+    if (xmldb_validate_db(target) < 0){
+	cprintf(cbret, "<rpc-reply><rpc-error>"
+		"<error-tag>invalid-value</error-tag>"
+		"<error-type>protocol</error-type>"
+		"<error-severity>error</error-severity>"
+		"<error-message>No such database: %s</error-message>"
+		"</rpc-error></rpc-reply>", target);
 	goto ok;
     }
     /* Check if target locked by other client */
@@ -556,7 +614,6 @@ from_client_copy_config(clicon_handle h,
 		piddb);
 	goto ok;
     }
-
     if (xmldb_copy(h, source, target) < 0){
 	cprintf(cbret, "<rpc-reply><rpc-error>"
 		"<error-tag>operation-failed</error-tag>"
@@ -601,6 +658,16 @@ from_client_delete_config(clicon_handle h,
 		"</rpc-error></rpc-reply>");
 	goto ok;
     }
+    if (xmldb_validate_db(target) < 0){
+	cprintf(cbret, "<rpc-reply><rpc-error>"
+		"<error-tag>invalid-value</error-tag>"
+		"<error-type>protocol</error-type>"
+		"<error-severity>error</error-severity>"
+		"<error-message>No such database: %s</error-message>"
+		"</rpc-error></rpc-reply>", target);
+	goto ok;
+    }
+
     /* Check if target locked by other client */
     piddb = xmldb_islocked(h, target);
     if (piddb && mypid != piddb){