+++ example/routing_cli.c
@@ -81,8 +81,9 @@ mycallback(clicon_handle h, cvec *cvv, cvec *argv)
 
     /* Show eth0 interfaces config using XPATH */
     if (clicon_rpc_get_config(h, "running","/interfaces/interface[name=eth0]",
-			      &xret) < 0)
+			      0, &xret) < 0)
 	goto done;
+
     xml_print(stdout, xret);
     retval = 0;
  done: