+++ lib/clixon/clixon_xml_db.h
@@ -75,12 +75,10 @@ typedef int (xmldb_getopt_t)(xmldb_handle xh, char *optname, void **value);
 typedef int (xmldb_setopt_t)(xmldb_handle xh, char *optname, void *value);
 
 /* Type of xmldb get function */
-typedef int (xmldb_get_t)(xmldb_handle xh, char *db, char *xpath,
-	   cxobj **xtop, cxobj ***xvec, size_t *xlen);
+typedef int (xmldb_get_t)(xmldb_handle xh, char *db, char *xpath, cxobj **xtop);
 
 /* Type of xmldb put function */
-typedef int (xmldb_put_t)(xmldb_handle xh, char *db, enum operation_type op, 
-			  cxobj *xt);
+typedef int (xmldb_put_t)(xmldb_handle xh, char *db, enum operation_type op, cxobj *xt);
 
 /* Type of xmldb copy function */
 typedef int (xmldb_copy_t)(xmldb_handle xh, char *from, char *to);
@@ -135,12 +133,12 @@ struct xmldb_api{
 int xmldb_plugin_load(clicon_handle h, char *filename);
 int xmldb_plugin_unload(clicon_handle h);
 
+int xmldb_validate_db(char *db);
 int xmldb_connect(clicon_handle h);
 int xmldb_disconnect(clicon_handle h);
 int xmldb_getopt(clicon_handle h, char *optname, void **value);
 int xmldb_setopt(clicon_handle h, char *optname, void *value);
-int xmldb_get(clicon_handle h, char *db, char *xpath,
-	      cxobj **xtop, cxobj ***xvec, size_t *xlen);
+int xmldb_get(clicon_handle h, char *db, char *xpath, cxobj **xtop);
 int xmldb_put(clicon_handle h, char *db, enum operation_type op, cxobj *xt);
 int xmldb_copy(clicon_handle h, char *from, char *to);
 int xmldb_lock(clicon_handle h, char *db, int pid);